backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]
pydantic==2.5.3
httpx==0.27.0
beautifulsoup4
python-multipart
jinja2 
python-dotenv
playwright            
huggingface_hub           
